re touch

+++ app/api/comment_routes.py
@@ -1,18 +1,20 @@
 from flask import Blueprint, jsonify
 from flask.globals import request
 from flask_login import login_required, current_user
-from app.models import Comment
+from app.models import Comment, Player
 from app import db
 
 comment_routes = Blueprint("comments", __name__)
 
 
-@comment_routes.route('/')
-#@login_required
-def get_comment():
-    comment = Comment.query.all
-    return comment.to_dict()
 
+@comment_routes.route('/<int:player_id>')
+#@login_required
+def get_comment(player_id):
+
+    player = Player.query.get(player_id)
+    return {comment.id:comment.to_dict() for comment in player.comments}
+    
 
 @comment_routes.route('/', methods=['post'])
 #@login_required
@@ -20,7 +22,10 @@ def post_comment():
 
 
     data = request.json
-    new = Comment(**data)
+    new = Comment(user = current_user,
+                  player_id  = data['playerId'],
+                  comment = data['postNewComment']   
+    )
 
     db.session.add(new)
     db.session.commit()
@@ -30,7 +35,8 @@ def post_comment():
 #@login_required
 def edit_comment(id):
     new = Comment.query.get(id)
-
+    if new.user_id != current_user.id:
+        return {}
     comment = request.json['comment']
     new.comment = comment
     
@@ -41,6 +47,8 @@ def edit_comment(id):
 #@login_required
 def delete_comment(id):
     comment = Comment.query.get(id)
+    if comment.user_id != current_user.id:
+        return {}
     db.session.delete(comment)
     db.session.commit()
-    return comment.to_dict()
+    return {}
